requirements.txt
Flask
requests
python-dotenv
gunicorn
google-generativeai
pytest
